+++ src/Compute/Compute/Strategies/ComputeRp/VirtualMachineStrategy.cs
@@ -72,9 +72,9 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             int? vCPUsPerCore = null,
             string imageReferenceId = null,
             Dictionary<string, List<string>> auxAuthHeader = null,
+            string diskControllerType = null,
             Microsoft.Azure.Management.Compute.Models.ExtendedLocation extendedLocation = null
             )
-
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
                 name: name,
@@ -116,7 +116,8 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                                 Id = imageReferenceId
                             },
                             DataDisks = DataDiskStrategy.CreateDataDisks(
-                                imageAndOsType?.DataDiskLuns, dataDisks, dataDiskDeleteOption)
+                                imageAndOsType?.DataDiskLuns, dataDisks, dataDiskDeleteOption),
+                            DiskControllerType = diskControllerType
                         },
                         AvailabilitySet = engine.GetReference(availabilitySet),
                         Zones = zones,