+++ src/Migrate/exports/New-AzMigrateNicMapping.ps1
@@ -61,7 +61,19 @@ param(
     [Microsoft.Azure.PowerShell.Cmdlets.Migrate.Category('Path')]
     [System.String]
     # Specifies the IP within the destination subnet to be used for the NIC.
-    ${TargetNicIP}
+    ${TargetNicIP},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.Migrate.Category('Path')]
+    [System.String]
+    # Specifies the Subnet name for the NIC in the destination Virtual Network to which the server needs to be test migrated.
+    ${TestNicSubnet},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.Migrate.Category('Path')]
+    [System.String]
+    # Specifies the IP within the destination test subnet to be used for the NIC.
+    ${TestNicIP}
 )
 
 begin {