+++ src/DataProtection/Changelog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed list parameter set for `Get-AzDataProtectionBackupVault`
+
+## Version 0.5.0
 * Added support for automatic assignment of permissions for configure backup for DPP workloads - AzureDisk, AzureBlob, AzureDatabaseForPostgreSQL using Set-AzDataProtectionMSIPermission cmdlet
 * Added support for adding custom tags for new backup instance using New-AzDataProtectionBackupInstance cmdlet